requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
flask==3.0.2
flask-cors==4.0.0 